FIRE-35011 - Initial check in
Added new value to track the sDesiredDiscordBias prior to it changing, to be able to see which direction the bias is moving.
Added time stamps to track when the system enters low memory and when it recovers back to normal.
Added flag to determine if a texture was sized down and used to delay the re-upsizing of the texture for a period of time to prevent issues of sudden ramp of requests for high res textures that were just downsized a few seconds ago.

This resolves the issue of the sudden surge of CPU usage and GPU usage and smoothed it out. There are other things that can be looked at to resolve the high resource usage of the viewer.

Another check in will come to fix up the code and clean it up more as there are lots of minor things that was tried, but have left over code that should be refactored and cleaned up.