Removed the particle count for validation as it is no longer needed.
As we now have a fixed pool of Particles, we no longer need to worry about allocating too many particles.
Repurposed the sParticleCount to be used to index into the vector particles.
Added proper initialization of the particle pool (we have a vector of pointers, so default constructor just turns them all NULL.
Will do tests with various particle examples.
Started to add flag to swap between both methods of handling particle memory.
This is not a final solution as a better particle system design would use large vertex and color arrays for use with VBO