Fixed up toggle to turn on the new Particle pool
Fixed up the particle pool flag so that all the features work correctly.
Note that I am using sParticleCount2 to track the old particles and disabled the use of it for checking to see if we go over.

With the new pool, it is impossible to go over the limit.