Added new features such as Atomic Fetch Texture worker state to speed up threading
Added new features such as Atomic Fetch Texture worker state
Changed the check for allowing access to the raws during saving to cache to after as there is conditions where the raw would not cached if they were modified by the rest of the engine.
Added 2 levels of low res textures to use.